문제 2 수정 및 문제 3 풀이

+++ src/chapter2/OperationAssignment2.java
@@ -7,7 +7,7 @@ public class OperationAssignment2 {
         int num1 = 7;
         int num2 = 2;
 
-        double result = (double) num1 / (double) num2;
+        double result = (double) num1 / num2;
         System.out.println(result);
     }
 }
